requirements.txt
Pillow
numpy
skimage